Show Error from console on GUI

+++ PoGo.NecroBot.Logic/MultiAccountManager.cs
@@ -112,10 +112,13 @@ namespace PoGo.NecroBot.Logic
             
             // Backup old config file.
             long ts = DateTime.UtcNow.ToUnixTime(); // Add timestamp to avoid file conflicts
-            string backupPath = $"accounts-{schemaVersion}-{ts}.backup.db";
-            Logging.Logger.Write($"Backing up accounts.db to: {backupPath}", LogLevel.Info);
-            File.Copy("accounts.db", backupPath);
+            if (File.Exists("accounts.db"))
+            {
+                string backupPath = $"accounts-{schemaVersion}-{ts}.backup.db";
+                Logging.Logger.Write($"Backing up accounts.db to: {backupPath}", LogLevel.Info);
             
+                File.Copy("accounts.db", backupPath);
+            }
             // Add future schema migrations below.
             int version;
             for (version = schemaVersion; version < UpdateConfig.CURRENT_SCHEMA_VERSION; version++)
